SYS-2825 Move tests into respective files
