+++ pallets/parachain-staking/src/tests/schedule_unbond_tests.rs
@@ -3,14 +3,14 @@
 #![cfg(test)]
 
 use crate::{
-    assert_event_emitted, encode_signed_execute_nomination_request_params,
+    assert_event_emitted, assert_last_event, encode_signed_execute_nomination_request_params,
     encode_signed_schedule_candidate_unbond_params, encode_signed_schedule_nominator_unbond_params,
     mock::{
-        build_proof, roll_to_era_begin, sign, AccountId, AvnProxy, Call as MockCall, ExtBuilder,
-        MinNominationPerCollator, Origin, ParachainStaking, Signature, Staker, System, Test,
-        TestAccount,
+        build_proof, roll_to, roll_to_era_begin, sign, AccountId, AvnProxy, Call as MockCall,
+        Event as MetaEvent, ExtBuilder, MinNominationPerCollator, Origin, ParachainStaking,
+        Signature, Staker, System, Test, TestAccount,
     },
-    Config, Error, Event, Proof,
+    Bond, Config, Error, Event, NominationAction, Proof, ScheduledRequest,
 };
 use frame_support::{assert_noop, assert_ok, error::BadOrigin};
 use frame_system::{self as system, RawOrigin};
@@ -773,4 +773,741 @@ mod signed_execute_nomination_request {
                 });
         }
     }
+}
+
+// NOMINATOR BOND LESS
+
+#[test]
+fn nominator_unbond_event_emits_correctly() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                5
+            ));
+            assert_last_event!(MetaEvent::ParachainStaking(Event::NominationDecreaseScheduled {
+                nominator: account_id_2,
+                candidate: account_id,
+                amount_to_decrease: 5,
+                execute_era: 3,
+            }));
+        });
+}
+
+#[test]
+fn nominator_unbond_updates_nominator_state() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                5
+            ));
+            let state = ParachainStaking::nomination_scheduled_requests(&account_id);
+            assert_eq!(
+                state,
+                vec![ScheduledRequest {
+                    nominator: account_id_2,
+                    when_executable: 3,
+                    action: NominationAction::Decrease(5),
+                }],
+            );
+        });
+}
+
+#[test]
+fn nominator_not_allowed_to_unbond_if_leaving() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 15)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_leave_nominators(Origin::signed(account_id_2)));
+            assert_noop!(
+                ParachainStaking::schedule_nominator_unbond(
+                    Origin::signed(account_id_2),
+                    account_id,
+                    1
+                ),
+                <Error<Test>>::PendingNominationRequestAlreadyExists,
+            );
+        });
+}
+
+#[test]
+fn cannot_nominator_unbond_if_revoking() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    let account_id_3 = to_acc_id(3u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 25), (account_id_3, 20)])
+        .with_candidates(vec![(account_id, 30), (account_id_3, 20)])
+        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_revoke_nomination(
+                Origin::signed(account_id_2),
+                account_id
+            ));
+            assert_noop!(
+                ParachainStaking::schedule_nominator_unbond(
+                    Origin::signed(account_id_2),
+                    account_id,
+                    1
+                ),
+                Error::<Test>::PendingNominationRequestAlreadyExists
+            );
+        });
+}
+
+#[test]
+fn cannot_nominator_unbond_if_not_nominator() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default().build().execute_with(|| {
+        assert_noop!(
+            ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                5
+            ),
+            Error::<Test>::NominatorDNE
+        );
+    });
+}
+
+#[test]
+fn cannot_nominator_unbond_if_candidate_dne() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_noop!(
+                ParachainStaking::schedule_nominator_unbond(
+                    Origin::signed(account_id_2),
+                    to_acc_id(3),
+                    5
+                ),
+                Error::<Test>::NominationDNE
+            );
+        });
+}
+
+#[test]
+fn cannot_nominator_unbond_if_nomination_dne() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    let account_id_3 = to_acc_id(3u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 10), (account_id_3, 30)])
+        .with_candidates(vec![(account_id, 30), (account_id_3, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_noop!(
+                ParachainStaking::schedule_nominator_unbond(
+                    Origin::signed(account_id_2),
+                    account_id_3,
+                    5
+                ),
+                Error::<Test>::NominationDNE
+            );
+        });
+}
+
+#[test]
+fn cannot_nominator_unbond_below_min_collator_stk() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_noop!(
+                ParachainStaking::schedule_nominator_unbond(
+                    Origin::signed(account_id_2),
+                    account_id,
+                    6
+                ),
+                Error::<Test>::NominatorBondBelowMin
+            );
+        });
+}
+
+#[test]
+fn cannot_nominator_unbond_more_than_total_nomination() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_noop!(
+                ParachainStaking::schedule_nominator_unbond(
+                    Origin::signed(account_id_2),
+                    account_id,
+                    11
+                ),
+                Error::<Test>::NominatorBondBelowMin
+            );
+        });
+}
+
+#[test]
+fn cannot_nominator_unbond_below_min_nomination() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    let account_id_3 = to_acc_id(3u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 20), (account_id_3, 30)])
+        .with_candidates(vec![(account_id, 30), (account_id_3, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
+        .build()
+        .execute_with(|| {
+            assert_noop!(
+                ParachainStaking::schedule_nominator_unbond(
+                    Origin::signed(account_id_2),
+                    account_id,
+                    8
+                ),
+                Error::<Test>::NominationBelowMin
+            );
+        });
+}
+
+// SCHEDULE CANDIDATE BOND LESS
+
+#[test]
+fn schedule_candidate_unbond_event_emits_correctly() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
+            assert_last_event!(MetaEvent::ParachainStaking(Event::CandidateBondLessRequested {
+                candidate: account_id,
+                amount_to_decrease: 10,
+                execute_era: 3,
+            }));
+        });
+}
+
+#[test]
+fn cannot_schedule_candidate_unbond_if_request_exists() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 5));
+            assert_noop!(
+                ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 5),
+                Error::<Test>::PendingCandidateRequestAlreadyExists
+            );
+        });
+}
+
+#[test]
+fn cannot_schedule_candidate_unbond_if_not_candidate() {
+    ExtBuilder::default().build().execute_with(|| {
+        assert_noop!(
+            ParachainStaking::schedule_candidate_unbond(Origin::signed(to_acc_id(6)), 50),
+            Error::<Test>::CandidateDNE
+        );
+    });
+}
+
+#[test]
+fn cannot_schedule_candidate_unbond_if_new_total_below_min_candidate_stk() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            assert_noop!(
+                ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 21),
+                Error::<Test>::CandidateBondBelowMin
+            );
+        });
+}
+
+#[test]
+fn can_schedule_candidate_unbond_if_leaving_candidates() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_leave_candidates(Origin::signed(account_id), 1));
+            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
+        });
+}
+
+#[test]
+fn cannot_schedule_candidate_unbond_if_exited_candidates() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_leave_candidates(Origin::signed(account_id), 1));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_leave_candidates(
+                Origin::signed(account_id),
+                account_id,
+                0
+            ));
+            assert_noop!(
+                ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10),
+                Error::<Test>::CandidateDNE
+            );
+        });
+}
+// EXECUTE CANDUDATE BOND LESS REQUEST
+
+#[test]
+fn execute_candidate_unbond_emits_correct_event() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 50)])
+        .with_candidates(vec![(account_id, 50)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 30));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_candidate_unbond(
+                Origin::signed(account_id),
+                account_id
+            ));
+            assert_last_event!(MetaEvent::ParachainStaking(Event::CandidateBondedLess {
+                candidate: account_id,
+                amount: 30,
+                new_bond: 20
+            }));
+        });
+}
+
+#[test]
+fn execute_candidate_unbond_unreserves_balance() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            assert_eq!(ParachainStaking::get_collator_stakable_free_balance(&account_id), 0);
+            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_candidate_unbond(
+                Origin::signed(account_id),
+                account_id
+            ));
+            assert_eq!(ParachainStaking::get_collator_stakable_free_balance(&account_id), 10);
+        });
+}
+
+#[test]
+fn execute_candidate_unbond_decreases_total() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            let mut total = ParachainStaking::total();
+            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_candidate_unbond(
+                Origin::signed(account_id),
+                account_id
+            ));
+            total -= 10;
+            assert_eq!(ParachainStaking::total(), total);
+        });
+}
+
+#[test]
+fn execute_candidate_unbond_updates_candidate_state() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            let candidate_state =
+                ParachainStaking::candidate_info(account_id).expect("updated => exists");
+            assert_eq!(candidate_state.bond, 30);
+            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_candidate_unbond(
+                Origin::signed(account_id),
+                account_id
+            ));
+            let candidate_state =
+                ParachainStaking::candidate_info(account_id).expect("updated => exists");
+            assert_eq!(candidate_state.bond, 20);
+        });
+}
+
+#[test]
+fn execute_candidate_unbond_updates_candidate_pool() {
+    let account_id = to_acc_id(1u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30)])
+        .with_candidates(vec![(account_id, 30)])
+        .build()
+        .execute_with(|| {
+            assert_eq!(ParachainStaking::candidate_pool().0[0].owner, account_id);
+            assert_eq!(ParachainStaking::candidate_pool().0[0].amount, 30);
+            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_candidate_unbond(
+                Origin::signed(account_id),
+                account_id
+            ));
+            assert_eq!(ParachainStaking::candidate_pool().0[0].owner, account_id);
+            assert_eq!(ParachainStaking::candidate_pool().0[0].amount, 20);
+        });
+}
+
+// 2. EXECUTE NOMINATOR BOND LESS
+
+#[test]
+fn execute_nominator_unbond_unreserves_balance() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_eq!(ParachainStaking::get_nominator_stakable_free_balance(&account_id_2), 0);
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                5
+            ));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_nomination_request(
+                Origin::signed(account_id_2),
+                account_id_2,
+                account_id
+            ));
+            assert_eq!(ParachainStaking::get_nominator_stakable_free_balance(&account_id_2), 5);
+        });
+}
+
+#[test]
+fn execute_nominator_unbond_decreases_total_staked() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_eq!(ParachainStaking::total(), 40);
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                5
+            ));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_nomination_request(
+                Origin::signed(account_id_2),
+                account_id_2,
+                account_id
+            ));
+            assert_eq!(ParachainStaking::total(), 35);
+        });
+}
+
+#[test]
+fn execute_nominator_unbond_updates_nominator_state() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 15)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_eq!(
+                ParachainStaking::nominator_state(account_id_2).expect("exists").total(),
+                10
+            );
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                5
+            ));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_nomination_request(
+                Origin::signed(account_id_2),
+                account_id_2,
+                account_id
+            ));
+            assert_eq!(ParachainStaking::nominator_state(account_id_2).expect("exists").total(), 5);
+        });
+}
+
+#[test]
+fn execute_nominator_unbond_updates_candidate_state() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 15)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 10)])
+        .build()
+        .execute_with(|| {
+            assert_eq!(
+                ParachainStaking::top_nominations(account_id).unwrap().nominations[0].owner,
+                account_id_2
+            );
+            assert_eq!(
+                ParachainStaking::top_nominations(account_id).unwrap().nominations[0].amount,
+                10
+            );
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                5
+            ));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_nomination_request(
+                Origin::signed(account_id_2),
+                account_id_2,
+                account_id
+            ));
+            assert_eq!(
+                ParachainStaking::top_nominations(account_id).unwrap().nominations[0].owner,
+                account_id_2
+            );
+            assert_eq!(
+                ParachainStaking::top_nominations(account_id).unwrap().nominations[0].amount,
+                5
+            );
+        });
+}
+
+#[test]
+fn execute_nominator_unbond_updates_just_bottom_nominations() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![
+            (account_id, 20),
+            (account_id_2, 10),
+            (to_acc_id(3), 11),
+            (to_acc_id(4), 12),
+            (to_acc_id(5), 14),
+            (to_acc_id(6), 15),
+        ])
+        .with_candidates(vec![(account_id, 20)])
+        .with_nominations(vec![
+            (account_id_2, account_id, 10),
+            (to_acc_id(3), account_id, 11),
+            (to_acc_id(4), account_id, 12),
+            (to_acc_id(5), account_id, 14),
+            (to_acc_id(6), account_id, 15),
+        ])
+        .build()
+        .execute_with(|| {
+            let pre_call_candidate_info =
+                ParachainStaking::candidate_info(&account_id).expect("nominated by all so exists");
+            let pre_call_top_nominations =
+                ParachainStaking::top_nominations(&account_id).expect("nominated by all so exists");
+            let pre_call_bottom_nominations = ParachainStaking::bottom_nominations(&account_id)
+                .expect("nominated by all so exists");
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                2
+            ));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_nomination_request(
+                Origin::signed(account_id_2),
+                account_id_2,
+                account_id
+            ));
+            let post_call_candidate_info =
+                ParachainStaking::candidate_info(&account_id).expect("nominated by all so exists");
+            let post_call_top_nominations =
+                ParachainStaking::top_nominations(&account_id).expect("nominated by all so exists");
+            let post_call_bottom_nominations = ParachainStaking::bottom_nominations(&account_id)
+                .expect("nominated by all so exists");
+            let mut not_equal = false;
+            for Bond { owner, amount } in pre_call_bottom_nominations.nominations {
+                for Bond { owner: post_owner, amount: post_amount } in
+                    &post_call_bottom_nominations.nominations
+                {
+                    if &owner == post_owner {
+                        if &amount != post_amount {
+                            not_equal = true;
+                            break
+                        }
+                    }
+                }
+            }
+            assert!(not_equal);
+            let mut equal = true;
+            for Bond { owner, amount } in pre_call_top_nominations.nominations {
+                for Bond { owner: post_owner, amount: post_amount } in
+                    &post_call_top_nominations.nominations
+                {
+                    if &owner == post_owner {
+                        if &amount != post_amount {
+                            equal = false;
+                            break
+                        }
+                    }
+                }
+            }
+            assert!(equal);
+            assert_eq!(
+                pre_call_candidate_info.total_counted,
+                post_call_candidate_info.total_counted
+            );
+        });
+}
+
+#[test]
+fn execute_nominator_unbond_does_not_delete_bottom_nominations() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    let account_id_6 = to_acc_id(6u64);
+    ExtBuilder::default()
+        .with_balances(vec![
+            (account_id, 20),
+            (account_id_2, 10),
+            (to_acc_id(3), 11),
+            (to_acc_id(4), 12),
+            (to_acc_id(5), 14),
+            (account_id_6, 15),
+        ])
+        .with_candidates(vec![(account_id, 20)])
+        .with_nominations(vec![
+            (account_id_2, account_id, 10),
+            (to_acc_id(3), account_id, 11),
+            (to_acc_id(4), account_id, 12),
+            (to_acc_id(5), account_id, 14),
+            (account_id_6, account_id, 15),
+        ])
+        .build()
+        .execute_with(|| {
+            let pre_call_candidate_info =
+                ParachainStaking::candidate_info(&account_id).expect("nominated by all so exists");
+            let pre_call_top_nominations =
+                ParachainStaking::top_nominations(&account_id).expect("nominated by all so exists");
+            let pre_call_bottom_nominations = ParachainStaking::bottom_nominations(&account_id)
+                .expect("nominated by all so exists");
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_6),
+                account_id,
+                4
+            ));
+            roll_to(10);
+            assert_ok!(ParachainStaking::execute_nomination_request(
+                Origin::signed(account_id_6),
+                account_id_6,
+                account_id
+            ));
+            let post_call_candidate_info =
+                ParachainStaking::candidate_info(&account_id).expect("nominated by all so exists");
+            let post_call_top_nominations =
+                ParachainStaking::top_nominations(&account_id).expect("nominated by all so exists");
+            let post_call_bottom_nominations = ParachainStaking::bottom_nominations(&account_id)
+                .expect("nominated by all so exists");
+            let mut equal = true;
+            for Bond { owner, amount } in pre_call_bottom_nominations.nominations {
+                for Bond { owner: post_owner, amount: post_amount } in
+                    &post_call_bottom_nominations.nominations
+                {
+                    if &owner == post_owner {
+                        if &amount != post_amount {
+                            equal = false;
+                            break
+                        }
+                    }
+                }
+            }
+            assert!(equal);
+            let mut not_equal = false;
+            for Bond { owner, amount } in pre_call_top_nominations.nominations {
+                for Bond { owner: post_owner, amount: post_amount } in
+                    &post_call_top_nominations.nominations
+                {
+                    if &owner == post_owner {
+                        if &amount != post_amount {
+                            not_equal = true;
+                            break
+                        }
+                    }
+                }
+            }
+            assert!(not_equal);
+            assert_eq!(
+                pre_call_candidate_info.total_counted - 4,
+                post_call_candidate_info.total_counted
+            );
+        });
+}
+
+#[test]
+fn can_execute_nominator_unbond_for_leaving_candidate() {
+    let account_id = to_acc_id(1u64);
+    let account_id_2 = to_acc_id(2u64);
+    ExtBuilder::default()
+        .with_balances(vec![(account_id, 30), (account_id_2, 15)])
+        .with_candidates(vec![(account_id, 30)])
+        .with_nominations(vec![(account_id_2, account_id, 15)])
+        .build()
+        .execute_with(|| {
+            assert_ok!(ParachainStaking::schedule_leave_candidates(Origin::signed(account_id), 1));
+            assert_ok!(ParachainStaking::schedule_nominator_unbond(
+                Origin::signed(account_id_2),
+                account_id,
+                5
+            ));
+            roll_to(10);
+            // can execute bond more nomination request for leaving candidate
+            assert_ok!(ParachainStaking::execute_nomination_request(
+                Origin::signed(account_id_2),
+                account_id_2,
+                account_id
+            ));
+        });
 }
