+++ pallets/parachain-staking/src/tests/tests.rs
@@ -1042,208 +1042,6 @@ fn cannot_go_online_if_leaving() {
         });
 }
 
-// SCHEDULE CANDIDATE BOND LESS
-
-#[test]
-fn schedule_candidate_unbond_event_emits_correctly() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
-            assert_last_event!(MetaEvent::ParachainStaking(Event::CandidateBondLessRequested {
-                candidate: account_id,
-                amount_to_decrease: 10,
-                execute_era: 3,
-            }));
-        });
-}
-
-#[test]
-fn cannot_schedule_candidate_unbond_if_request_exists() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 5));
-            assert_noop!(
-                ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 5),
-                Error::<Test>::PendingCandidateRequestAlreadyExists
-            );
-        });
-}
-
-#[test]
-fn cannot_schedule_candidate_unbond_if_not_candidate() {
-    ExtBuilder::default().build().execute_with(|| {
-        assert_noop!(
-            ParachainStaking::schedule_candidate_unbond(Origin::signed(to_acc_id(6)), 50),
-            Error::<Test>::CandidateDNE
-        );
-    });
-}
-
-#[test]
-fn cannot_schedule_candidate_unbond_if_new_total_below_min_candidate_stk() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            assert_noop!(
-                ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 21),
-                Error::<Test>::CandidateBondBelowMin
-            );
-        });
-}
-
-#[test]
-fn can_schedule_candidate_unbond_if_leaving_candidates() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_leave_candidates(Origin::signed(account_id), 1));
-            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
-        });
-}
-
-#[test]
-fn cannot_schedule_candidate_unbond_if_exited_candidates() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_leave_candidates(Origin::signed(account_id), 1));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_leave_candidates(
-                Origin::signed(account_id),
-                account_id,
-                0
-            ));
-            assert_noop!(
-                ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10),
-                Error::<Test>::CandidateDNE
-            );
-        });
-}
-
-// 2. EXECUTE BOND LESS REQUEST
-
-#[test]
-fn execute_candidate_unbond_emits_correct_event() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 50)])
-        .with_candidates(vec![(account_id, 50)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 30));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_candidate_unbond(
-                Origin::signed(account_id),
-                account_id
-            ));
-            assert_last_event!(MetaEvent::ParachainStaking(Event::CandidateBondedLess {
-                candidate: account_id,
-                amount: 30,
-                new_bond: 20
-            }));
-        });
-}
-
-#[test]
-fn execute_candidate_unbond_unreserves_balance() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(ParachainStaking::get_collator_stakable_free_balance(&account_id), 0);
-            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_candidate_unbond(
-                Origin::signed(account_id),
-                account_id
-            ));
-            assert_eq!(ParachainStaking::get_collator_stakable_free_balance(&account_id), 10);
-        });
-}
-
-#[test]
-fn execute_candidate_unbond_decreases_total() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            let mut total = ParachainStaking::total();
-            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_candidate_unbond(
-                Origin::signed(account_id),
-                account_id
-            ));
-            total -= 10;
-            assert_eq!(ParachainStaking::total(), total);
-        });
-}
-
-#[test]
-fn execute_candidate_unbond_updates_candidate_state() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            let candidate_state =
-                ParachainStaking::candidate_info(account_id).expect("updated => exists");
-            assert_eq!(candidate_state.bond, 30);
-            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_candidate_unbond(
-                Origin::signed(account_id),
-                account_id
-            ));
-            let candidate_state =
-                ParachainStaking::candidate_info(account_id).expect("updated => exists");
-            assert_eq!(candidate_state.bond, 20);
-        });
-}
-
-#[test]
-fn execute_candidate_unbond_updates_candidate_pool() {
-    let account_id = to_acc_id(1u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30)])
-        .with_candidates(vec![(account_id, 30)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(ParachainStaking::candidate_pool().0[0].owner, account_id);
-            assert_eq!(ParachainStaking::candidate_pool().0[0].amount, 30);
-            assert_ok!(ParachainStaking::schedule_candidate_unbond(Origin::signed(account_id), 10));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_candidate_unbond(
-                Origin::signed(account_id),
-                account_id
-            ));
-            assert_eq!(ParachainStaking::candidate_pool().0[0].owner, account_id);
-            assert_eq!(ParachainStaking::candidate_pool().0[0].amount, 20);
-        });
-}
-
 // CANCEL CANDIDATE BOND LESS REQUEST
 
 #[test]
@@ -1741,1141 +1539,6 @@ fn cannot_cancel_leave_nominators_if_single_nomination_revoke_manually_cancelled
         });
 }
 
-// SCHEDULE REVOKE NOMINATION
-
-#[test]
-fn revoke_nomination_event_emits_correctly() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_3 = to_acc_id(3u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 20), (account_id_3, 30)])
-        .with_candidates(vec![(account_id, 30), (account_id_3, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            assert_last_event!(MetaEvent::ParachainStaking(Event::NominationRevocationScheduled {
-                era: 1,
-                nominator: account_id_2,
-                candidate: account_id,
-                scheduled_exit: 3,
-            }));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_event_emitted!(Event::NominatorLeftCandidate {
-                nominator: account_id_2,
-                candidate: account_id,
-                unstaked_amount: 10,
-                total_candidate_staked: 30
-            });
-        });
-}
-
-#[test]
-fn can_revoke_nomination_if_revoking_another_nomination() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_3 = to_acc_id(3u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 20), (account_id_3, 20)])
-        .with_candidates(vec![(account_id, 30), (account_id_3, 20)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            // this is an exit implicitly because last nomination revoked
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id_3
-            ));
-        });
-}
-
-#[test]
-fn nominator_not_allowed_revoke_if_already_leaving() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_3 = to_acc_id(3u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 20), (account_id_3, 20)])
-        .with_candidates(vec![(account_id, 30), (account_id_3, 20)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_leave_nominators(Origin::signed(account_id_2)));
-            assert_noop!(
-                ParachainStaking::schedule_revoke_nomination(
-                    Origin::signed(account_id_2),
-                    account_id_3
-                ),
-                <Error<Test>>::PendingNominationRequestAlreadyExists,
-            );
-        });
-}
-
-#[test]
-fn cannot_revoke_nomination_if_not_nominator() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default().build().execute_with(|| {
-        assert_noop!(
-            ParachainStaking::schedule_revoke_nomination(Origin::signed(account_id_2), account_id),
-            Error::<Test>::NominatorDNE
-        );
-    });
-}
-
-#[test]
-fn cannot_revoke_nomination_that_dne() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_noop!(
-                ParachainStaking::schedule_revoke_nomination(
-                    Origin::signed(account_id_2),
-                    to_acc_id(3)
-                ),
-                Error::<Test>::NominationDNE
-            );
-        });
-}
-
-#[test]
-// See `cannot_execute_revoke_nomination_below_min_nominator_stake` for where the "must be above
-// MinTotalNominatorStake" rule is now enforced.
-fn can_schedule_revoke_nomination_below_min_nominator_stake() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 20), (account_id_2, 8), (to_acc_id(3), 20)])
-        .with_candidates(vec![(account_id, 20), (to_acc_id(3), 20)])
-        .with_nominations(vec![(account_id_2, account_id, 5), (account_id_2, to_acc_id(3), 3)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-        });
-}
-
-// NOMINATOR BOND LESS
-
-#[test]
-fn nominator_unbond_event_emits_correctly() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                5
-            ));
-            assert_last_event!(MetaEvent::ParachainStaking(Event::NominationDecreaseScheduled {
-                nominator: account_id_2,
-                candidate: account_id,
-                amount_to_decrease: 5,
-                execute_era: 3,
-            }));
-        });
-}
-
-#[test]
-fn nominator_unbond_updates_nominator_state() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                5
-            ));
-            let state = ParachainStaking::nomination_scheduled_requests(&account_id);
-            assert_eq!(
-                state,
-                vec![ScheduledRequest {
-                    nominator: account_id_2,
-                    when_executable: 3,
-                    action: NominationAction::Decrease(5),
-                }],
-            );
-        });
-}
-
-#[test]
-fn nominator_not_allowed_to_unbond_if_leaving() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 15)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_leave_nominators(Origin::signed(account_id_2)));
-            assert_noop!(
-                ParachainStaking::schedule_nominator_unbond(
-                    Origin::signed(account_id_2),
-                    account_id,
-                    1
-                ),
-                <Error<Test>>::PendingNominationRequestAlreadyExists,
-            );
-        });
-}
-
-#[test]
-fn cannot_nominator_unbond_if_revoking() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_3 = to_acc_id(3u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 25), (account_id_3, 20)])
-        .with_candidates(vec![(account_id, 30), (account_id_3, 20)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            assert_noop!(
-                ParachainStaking::schedule_nominator_unbond(
-                    Origin::signed(account_id_2),
-                    account_id,
-                    1
-                ),
-                Error::<Test>::PendingNominationRequestAlreadyExists
-            );
-        });
-}
-
-#[test]
-fn cannot_nominator_unbond_if_not_nominator() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default().build().execute_with(|| {
-        assert_noop!(
-            ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                5
-            ),
-            Error::<Test>::NominatorDNE
-        );
-    });
-}
-
-#[test]
-fn cannot_nominator_unbond_if_candidate_dne() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_noop!(
-                ParachainStaking::schedule_nominator_unbond(
-                    Origin::signed(account_id_2),
-                    to_acc_id(3),
-                    5
-                ),
-                Error::<Test>::NominationDNE
-            );
-        });
-}
-
-#[test]
-fn cannot_nominator_unbond_if_nomination_dne() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_3 = to_acc_id(3u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10), (account_id_3, 30)])
-        .with_candidates(vec![(account_id, 30), (account_id_3, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_noop!(
-                ParachainStaking::schedule_nominator_unbond(
-                    Origin::signed(account_id_2),
-                    account_id_3,
-                    5
-                ),
-                Error::<Test>::NominationDNE
-            );
-        });
-}
-
-#[test]
-fn cannot_nominator_unbond_below_min_collator_stk() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_noop!(
-                ParachainStaking::schedule_nominator_unbond(
-                    Origin::signed(account_id_2),
-                    account_id,
-                    6
-                ),
-                Error::<Test>::NominatorBondBelowMin
-            );
-        });
-}
-
-#[test]
-fn cannot_nominator_unbond_more_than_total_nomination() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_noop!(
-                ParachainStaking::schedule_nominator_unbond(
-                    Origin::signed(account_id_2),
-                    account_id,
-                    11
-                ),
-                Error::<Test>::NominatorBondBelowMin
-            );
-        });
-}
-
-#[test]
-fn cannot_nominator_unbond_below_min_nomination() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_3 = to_acc_id(3u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 20), (account_id_3, 30)])
-        .with_candidates(vec![(account_id, 30), (account_id_3, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
-        .build()
-        .execute_with(|| {
-            assert_noop!(
-                ParachainStaking::schedule_nominator_unbond(
-                    Origin::signed(account_id_2),
-                    account_id,
-                    8
-                ),
-                Error::<Test>::NominationBelowMin
-            );
-        });
-}
-
-// EXECUTE PENDING NOMINATION REQUEST
-
-// 1. REVOKE NOMINATION
-
-#[test]
-fn execute_revoke_nomination_emits_exit_event_if_exit_happens() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    // last nomination is revocation
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_event_emitted!(Event::NominatorLeftCandidate {
-                nominator: account_id_2,
-                candidate: account_id,
-                unstaked_amount: 10,
-                total_candidate_staked: 30
-            });
-            assert_event_emitted!(Event::NominatorLeft {
-                nominator: account_id_2,
-                unstaked_amount: 10
-            });
-        });
-}
-
-#[test]
-fn cannot_execute_revoke_nomination_below_min_nominator_stake() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 20), (account_id_2, 8), (to_acc_id(3), 20)])
-        .with_candidates(vec![(account_id, 20), (to_acc_id(3), 20)])
-        .with_nominations(vec![(account_id_2, account_id, 5), (account_id_2, to_acc_id(3), 3)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_noop!(
-                ParachainStaking::execute_nomination_request(
-                    Origin::signed(account_id_2),
-                    account_id_2,
-                    account_id
-                ),
-                Error::<Test>::NominatorBondBelowMin
-            );
-            // but nominator can cancel the request and request to leave instead:
-            assert_ok!(ParachainStaking::cancel_nomination_request(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            assert_ok!(ParachainStaking::schedule_leave_nominators(Origin::signed(account_id_2)));
-            roll_to(20);
-            assert_ok!(ParachainStaking::execute_leave_nominators(
-                Origin::signed(account_id_2),
-                account_id_2,
-                2
-            ));
-        });
-}
-
-#[test]
-fn revoke_nomination_executes_exit_if_last_nomination() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    // last nomination is revocation
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_event_emitted!(Event::NominatorLeftCandidate {
-                nominator: account_id_2,
-                candidate: account_id,
-                unstaked_amount: 10,
-                total_candidate_staked: 30
-            });
-            assert_event_emitted!(Event::NominatorLeft {
-                nominator: account_id_2,
-                unstaked_amount: 10
-            });
-        });
-}
-
-#[test]
-fn execute_revoke_nomination_emits_correct_event() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 20), (to_acc_id(3), 30)])
-        .with_candidates(vec![(account_id, 30), (to_acc_id(3), 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, to_acc_id(3), 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_event_emitted!(Event::NominatorLeftCandidate {
-                nominator: account_id_2,
-                candidate: account_id,
-                unstaked_amount: 10,
-                total_candidate_staked: 30
-            });
-        });
-}
-
-#[test]
-fn execute_revoke_nomination_unreserves_balance() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(ParachainStaking::get_nominator_stakable_free_balance(&account_id_2), 0);
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_eq!(ParachainStaking::get_nominator_stakable_free_balance(&account_id_2), 10);
-        });
-}
-
-#[test]
-fn execute_revoke_nomination_adds_revocation_to_nominator_state() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 20), (to_acc_id(3), 20)])
-        .with_candidates(vec![(account_id, 30), (to_acc_id(3), 20)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, to_acc_id(3), 10)])
-        .build()
-        .execute_with(|| {
-            assert!(!ParachainStaking::nomination_scheduled_requests(&account_id)
-                .iter()
-                .any(|x| x.nominator == account_id_2));
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            assert!(ParachainStaking::nomination_scheduled_requests(&account_id)
-                .iter()
-                .any(|x| x.nominator == account_id_2));
-        });
-}
-
-#[test]
-fn execute_revoke_nomination_removes_revocation_from_nominator_state_upon_execution() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 20), (to_acc_id(3), 20)])
-        .with_candidates(vec![(account_id, 30), (to_acc_id(3), 20)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, to_acc_id(3), 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert!(!ParachainStaking::nomination_scheduled_requests(&account_id)
-                .iter()
-                .any(|x| x.nominator == account_id_2));
-        });
-}
-
-#[test]
-fn execute_revoke_nomination_removes_revocation_from_state_for_single_nomination_leave() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 20), (to_acc_id(3), 20)])
-        .with_candidates(vec![(account_id, 30), (to_acc_id(3), 20)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert!(
-                !ParachainStaking::nomination_scheduled_requests(&account_id)
-                    .iter()
-                    .any(|x| x.nominator == account_id_2),
-                "nomination was not removed"
-            );
-        });
-}
-
-#[test]
-fn execute_revoke_nomination_decreases_total_staked() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(ParachainStaking::total(), 40);
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_eq!(ParachainStaking::total(), 30);
-        });
-}
-
-#[test]
-fn execute_revoke_nomination_for_last_nomination_removes_nominator_state() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert!(ParachainStaking::nominator_state(account_id_2).is_some());
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            // this will be confusing for people
-            // if status is leaving, then execute_nomination_request works if last nomination
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert!(ParachainStaking::nominator_state(account_id_2).is_none());
-        });
-}
-
-#[test]
-fn execute_revoke_nomination_removes_nomination_from_candidate_state() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(
-                ParachainStaking::candidate_info(account_id).expect("exists").nomination_count,
-                1u32
-            );
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert!(ParachainStaking::candidate_info(account_id)
-                .expect("exists")
-                .nomination_count
-                .is_zero());
-        });
-}
-
-#[test]
-fn can_execute_revoke_nomination_for_leaving_candidate() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_leave_candidates(Origin::signed(account_id), 1));
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            // can execute nomination request for leaving candidate
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-        });
-}
-
-#[test]
-fn can_execute_leave_candidates_if_revoking_candidate() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_leave_candidates(Origin::signed(account_id), 1));
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            roll_to(10);
-            // revocation executes during execute leave candidates (callable by anyone)
-            assert_ok!(ParachainStaking::execute_leave_candidates(
-                Origin::signed(account_id),
-                account_id,
-                1
-            ));
-            assert!(!ParachainStaking::is_nominator(&account_id_2));
-            assert_eq!(Balances::reserved_balance(&account_id_2), 0);
-            assert_eq!(Balances::free_balance(&account_id_2), 10);
-        });
-}
-
-#[test]
-fn nominator_bond_more_after_revoke_nomination_does_not_effect_exit() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_3 = to_acc_id(3u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 30), (account_id_3, 30)])
-        .with_candidates(vec![(account_id, 30), (account_id_3, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            assert_ok!(ParachainStaking::bond_extra(
-                Origin::signed(account_id_2),
-                account_id_3,
-                10
-            ));
-            roll_to(100);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert!(ParachainStaking::is_nominator(&account_id_2));
-            assert_eq!(ParachainStaking::get_nominator_stakable_free_balance(&account_id_2), 10);
-        });
-}
-
-#[test]
-fn nominator_unbond_after_revoke_nomination_does_not_effect_exit() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_3 = to_acc_id(3u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 30), (account_id_3, 30)])
-        .with_candidates(vec![(account_id, 30), (account_id_3, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10), (account_id_2, account_id_3, 10)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_revoke_nomination(
-                Origin::signed(account_id_2),
-                account_id
-            ));
-            assert_last_event!(MetaEvent::ParachainStaking(Event::NominationRevocationScheduled {
-                era: 1,
-                nominator: account_id_2,
-                candidate: account_id,
-                scheduled_exit: 3,
-            }));
-            assert_noop!(
-                ParachainStaking::schedule_nominator_unbond(
-                    Origin::signed(account_id_2),
-                    account_id,
-                    2
-                ),
-                Error::<Test>::PendingNominationRequestAlreadyExists
-            );
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id_3,
-                2
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id_3
-            ));
-            assert_last_event!(MetaEvent::ParachainStaking(Event::NominationDecreased {
-                nominator: account_id_2,
-                candidate: account_id_3,
-                amount: 2,
-                in_top: true
-            }));
-            assert!(ParachainStaking::is_nominator(&account_id_2));
-            assert_eq!(ParachainStaking::get_nominator_stakable_free_balance(&account_id_2), 22);
-        });
-}
-
-// 2. EXECUTE BOND LESS
-
-#[test]
-fn execute_nominator_unbond_unreserves_balance() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(ParachainStaking::get_nominator_stakable_free_balance(&account_id_2), 0);
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                5
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_eq!(ParachainStaking::get_nominator_stakable_free_balance(&account_id_2), 5);
-        });
-}
-
-#[test]
-fn execute_nominator_unbond_decreases_total_staked() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 10)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(ParachainStaking::total(), 40);
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                5
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_eq!(ParachainStaking::total(), 35);
-        });
-}
-
-#[test]
-fn execute_nominator_unbond_updates_nominator_state() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 15)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(
-                ParachainStaking::nominator_state(account_id_2).expect("exists").total(),
-                10
-            );
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                5
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_eq!(ParachainStaking::nominator_state(account_id_2).expect("exists").total(), 5);
-        });
-}
-
-#[test]
-fn execute_nominator_unbond_updates_candidate_state() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 15)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 10)])
-        .build()
-        .execute_with(|| {
-            assert_eq!(
-                ParachainStaking::top_nominations(account_id).unwrap().nominations[0].owner,
-                account_id_2
-            );
-            assert_eq!(
-                ParachainStaking::top_nominations(account_id).unwrap().nominations[0].amount,
-                10
-            );
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                5
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            assert_eq!(
-                ParachainStaking::top_nominations(account_id).unwrap().nominations[0].owner,
-                account_id_2
-            );
-            assert_eq!(
-                ParachainStaking::top_nominations(account_id).unwrap().nominations[0].amount,
-                5
-            );
-        });
-}
-
-#[test]
-fn execute_nominator_unbond_updates_just_bottom_nominations() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![
-            (account_id, 20),
-            (account_id_2, 10),
-            (to_acc_id(3), 11),
-            (to_acc_id(4), 12),
-            (to_acc_id(5), 14),
-            (to_acc_id(6), 15),
-        ])
-        .with_candidates(vec![(account_id, 20)])
-        .with_nominations(vec![
-            (account_id_2, account_id, 10),
-            (to_acc_id(3), account_id, 11),
-            (to_acc_id(4), account_id, 12),
-            (to_acc_id(5), account_id, 14),
-            (to_acc_id(6), account_id, 15),
-        ])
-        .build()
-        .execute_with(|| {
-            let pre_call_candidate_info =
-                ParachainStaking::candidate_info(&account_id).expect("nominated by all so exists");
-            let pre_call_top_nominations =
-                ParachainStaking::top_nominations(&account_id).expect("nominated by all so exists");
-            let pre_call_bottom_nominations = ParachainStaking::bottom_nominations(&account_id)
-                .expect("nominated by all so exists");
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                2
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-            let post_call_candidate_info =
-                ParachainStaking::candidate_info(&account_id).expect("nominated by all so exists");
-            let post_call_top_nominations =
-                ParachainStaking::top_nominations(&account_id).expect("nominated by all so exists");
-            let post_call_bottom_nominations = ParachainStaking::bottom_nominations(&account_id)
-                .expect("nominated by all so exists");
-            let mut not_equal = false;
-            for Bond { owner, amount } in pre_call_bottom_nominations.nominations {
-                for Bond { owner: post_owner, amount: post_amount } in
-                    &post_call_bottom_nominations.nominations
-                {
-                    if &owner == post_owner {
-                        if &amount != post_amount {
-                            not_equal = true;
-                            break
-                        }
-                    }
-                }
-            }
-            assert!(not_equal);
-            let mut equal = true;
-            for Bond { owner, amount } in pre_call_top_nominations.nominations {
-                for Bond { owner: post_owner, amount: post_amount } in
-                    &post_call_top_nominations.nominations
-                {
-                    if &owner == post_owner {
-                        if &amount != post_amount {
-                            equal = false;
-                            break
-                        }
-                    }
-                }
-            }
-            assert!(equal);
-            assert_eq!(
-                pre_call_candidate_info.total_counted,
-                post_call_candidate_info.total_counted
-            );
-        });
-}
-
-#[test]
-fn execute_nominator_unbond_does_not_delete_bottom_nominations() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    let account_id_6 = to_acc_id(6u64);
-    ExtBuilder::default()
-        .with_balances(vec![
-            (account_id, 20),
-            (account_id_2, 10),
-            (to_acc_id(3), 11),
-            (to_acc_id(4), 12),
-            (to_acc_id(5), 14),
-            (account_id_6, 15),
-        ])
-        .with_candidates(vec![(account_id, 20)])
-        .with_nominations(vec![
-            (account_id_2, account_id, 10),
-            (to_acc_id(3), account_id, 11),
-            (to_acc_id(4), account_id, 12),
-            (to_acc_id(5), account_id, 14),
-            (account_id_6, account_id, 15),
-        ])
-        .build()
-        .execute_with(|| {
-            let pre_call_candidate_info =
-                ParachainStaking::candidate_info(&account_id).expect("nominated by all so exists");
-            let pre_call_top_nominations =
-                ParachainStaking::top_nominations(&account_id).expect("nominated by all so exists");
-            let pre_call_bottom_nominations = ParachainStaking::bottom_nominations(&account_id)
-                .expect("nominated by all so exists");
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_6),
-                account_id,
-                4
-            ));
-            roll_to(10);
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_6),
-                account_id_6,
-                account_id
-            ));
-            let post_call_candidate_info =
-                ParachainStaking::candidate_info(&account_id).expect("nominated by all so exists");
-            let post_call_top_nominations =
-                ParachainStaking::top_nominations(&account_id).expect("nominated by all so exists");
-            let post_call_bottom_nominations = ParachainStaking::bottom_nominations(&account_id)
-                .expect("nominated by all so exists");
-            let mut equal = true;
-            for Bond { owner, amount } in pre_call_bottom_nominations.nominations {
-                for Bond { owner: post_owner, amount: post_amount } in
-                    &post_call_bottom_nominations.nominations
-                {
-                    if &owner == post_owner {
-                        if &amount != post_amount {
-                            equal = false;
-                            break
-                        }
-                    }
-                }
-            }
-            assert!(equal);
-            let mut not_equal = false;
-            for Bond { owner, amount } in pre_call_top_nominations.nominations {
-                for Bond { owner: post_owner, amount: post_amount } in
-                    &post_call_top_nominations.nominations
-                {
-                    if &owner == post_owner {
-                        if &amount != post_amount {
-                            not_equal = true;
-                            break
-                        }
-                    }
-                }
-            }
-            assert!(not_equal);
-            assert_eq!(
-                pre_call_candidate_info.total_counted - 4,
-                post_call_candidate_info.total_counted
-            );
-        });
-}
-
-#[test]
-fn can_execute_nominator_unbond_for_leaving_candidate() {
-    let account_id = to_acc_id(1u64);
-    let account_id_2 = to_acc_id(2u64);
-    ExtBuilder::default()
-        .with_balances(vec![(account_id, 30), (account_id_2, 15)])
-        .with_candidates(vec![(account_id, 30)])
-        .with_nominations(vec![(account_id_2, account_id, 15)])
-        .build()
-        .execute_with(|| {
-            assert_ok!(ParachainStaking::schedule_leave_candidates(Origin::signed(account_id), 1));
-            assert_ok!(ParachainStaking::schedule_nominator_unbond(
-                Origin::signed(account_id_2),
-                account_id,
-                5
-            ));
-            roll_to(10);
-            // can execute bond more nomination request for leaving candidate
-            assert_ok!(ParachainStaking::execute_nomination_request(
-                Origin::signed(account_id_2),
-                account_id_2,
-                account_id
-            ));
-        });
-}
-
 // CANCEL PENDING NOMINATION REQUEST
 // 1. CANCEL REVOKE NOMINATION
 
@@ -3130,7 +1793,7 @@ fn nominator_schedule_revocation_total() {
 }
 
 #[test]
-fn parachain_bond_inflation_reserve_matches_config() {
+fn parachain_flow_reserve_matches_config() {
     let account_id = to_acc_id(1u64);
     let account_id_2 = to_acc_id(2u64);
     let account_id_3 = to_acc_id(3u64);
@@ -5092,7 +3755,7 @@ fn payouts_follow_nomination_changes() {
                 account_id_6,
                 10
             ));
-            // keep paying 6 (note: inflation is in terms of total issuance so that's why 1 is 21)
+            // keep paying 6
             let mut new2 = vec![
                 Event::NominatorExitScheduled {
                     era: 4,
